requirements.txt
flask
pandas
tensorflow
nltk
scikit-learn
selenium
google-generativeai
